requirements.txt
openai-clip
opencv-python
numpy
matplotlib
requests
tqdm
rarfile